refactor(rag): extract RRF fusion into rag/fusion.rs
Convert rag.rs to rag/mod.rs and move web_to_search_result, rrf_merge,
dedup_merged into fusion.rs. Inline rrf_merge tests relocated to
tests/rag_fusion_test.rs. Public path application::rag::rrf_merge preserved
via re-export. Behavior unchanged.
